Extract standardResult from pageResult module
